# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для интернет-магазина обуви",
        "title": "Shoes Store API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/backup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает полный бэкап базы данных и сохраняет его в папку проекта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backup"
                ],
                "summary": "Создание бэкапа базы данных",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BackupResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/backup/download/{filename}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Скачивает указанный файл бэкапа",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Admin Backup"
                ],
                "summary": "Скачать файл бэкапа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя файла бэкапа",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SQL файл бэкапа",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный формат файла",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Файл бэкапа не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/backup/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список всех доступных файлов бэкапа в папке проекта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backup"
                ],
                "summary": "Получение информации о бэкапах",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BackupListResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/backup/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Восстанавливает базу данных из указанного файла бэкапа",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backup"
                ],
                "summary": "Восстановить базу данных из бэкапа",
                "parameters": [
                    {
                        "type": "file",
                        "description": "SQL файл бэкапа",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BackupResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат файла",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/backup/{filename}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет указанный файл бэкапа из папки проекта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backup"
                ],
                "summary": "Удаление файла бэкапа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя файла бэкапа",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BackupDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат файла",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Файл бэкапа не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BackupDeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.BackupInfo": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "size_mb": {
                    "type": "number"
                }
            }
        },
        "models.BackupListResponse": {
            "type": "object",
            "properties": {
                "backup_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BackupInfo"
                    }
                },
                "desktop_path": {
                    "type": "string"
                },
                "total_files": {
                    "type": "integer"
                }
            }
        },
        "models.BackupResponse": {
            "type": "object",
            "properties": {
                "file_path": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}